moved cover_complex.py outside sklearn/
