--- src/python/gudhi/sklearn/__init__.py
@@ -1,3 +1,0 @@
-from .cover_complex import *
-
-__all__ = ["NerveComplex", "GraphInducedComplex", "MapperComplex"]
