+++ src/python/test/test_cover_complex.py
@@ -9,7 +9,7 @@
 """
 
 from gudhi import CoverComplex
-from gudhi.sklearn import MapperComplex, GraphInducedComplex, NerveComplex
+from gudhi import MapperComplex, GraphInducedComplex, NerveComplex
 import pytest
 import numpy as np
 from sklearn.cluster import AgglomerativeClustering
